extract hook to own file
